# src/BlitzRelay/Http/RelayHttpApi.cs
using BlitzRelay.Hosting;
using BlitzRelay.Networking;
using BlitzRelay.Protocol;
using BlitzRelay.Rooms;
using BlitzRelay.Serialization;
using Microsoft.AspNetCore.Builder;
using Microsoft.AspNetCore.Hosting;
using Microsoft.AspNetCore.Http;
using Microsoft.Extensions.DependencyInjection;
using System.Buffers;
using System.Globalization;
using System.Threading.RateLimiting;
using System.Security.Cryptography;
using System.Text;

namespace BlitzRelay.Http;

internal static class RelayHttpApi
{
	private const string PublicRoomRateLimitPolicy = "public-room";
	private const int MaximumMetadataEntries = 16;
	private const int MaximumMetadataKeyBytes = 64;
	private const int MaximumMetadataValueBytes = 2048;
	private const int MaximumMetadataTotalBytes = 8192;

	private static readonly HashSet<string> PublicMetadataKeys = new(StringComparer.OrdinalIgnoreCase)
	{
		"HostName",
		"Version",
		"Users",
		"MaxMembers",
		"DisconnectedFriends",
	};

	public static WebApplication Build(WebApplicationBuilder builder, RelayHostOptions relayHostOptions)
	{
		builder.WebHost.ConfigureKestrel(options => options.ListenAnyIP(relayHostOptions.HttpPort));

		builder.Services.ConfigureHttpJsonOptions(options => options.SerializerOptions.TypeInfoResolverChain.Insert(0, AppJsonSerializerContext.Default));

		builder.Services.AddRateLimiter(options =>
		{
			options.AddPolicy(PublicRoomRateLimitPolicy, context => RateLimitPartition.GetFixedWindowLimiter
			(
				context.Connection.RemoteIpAddress?.ToString() ?? "unknown",
				_ => new FixedWindowRateLimiterOptions
				{
					PermitLimit = 60,
					Window = TimeSpan.FromMinutes(1),
					QueueLimit = 0,
					AutoReplenishment = true,
				}
			));
		});

		builder.Services.AddCors(options =>
		{
			options.AddDefaultPolicy(configure =>
			{
				CorsConfiguration? cors = relayHostOptions.Cors;

				if (cors is null)
				{
					configure.AllowAnyOrigin()
							 .AllowAnyMethod()
							 .AllowAnyHeader();

					return;
				}

				if (cors.AllowedOrigins is { Length: > 0 } origins)
				{
					configure.WithOrigins(origins);

					if (cors.AllowCredentials) configure.AllowCredentials();
				}
				else
				{
					configure.AllowAnyOrigin();
				}

				if (cors.AllowedMethods is { Length: > 0 } methods)
				{
					configure.WithMethods(methods);
				}
				else
				{
					configure.AllowAnyMethod();
				}

				if (cors.AllowedHeaders is { Length: > 0 } headers)
				{
					configure.WithHeaders(headers);
				}
				else
				{
					configure.AllowAnyHeader();
				}
			});
		});

		WebApplication app = builder.Build();

		Server relayServer = app.Services.GetRequiredService<Server>();

		app.UseCors();

		app.UseRateLimiter();

		app.MapGet("/health", () => Results.Ok());

		app.MapPost("/rooms", (HttpContext context, CreateRoomRequest createRoomRequest) => CreateRoom(context, createRoomRequest, relayServer, relayHostOptions.HttpAdminTokenBytes));

		app.MapGet("/rooms", (HttpContext context) => GetRooms(context, relayServer, relayHostOptions.HttpAdminTokenBytes));

		app.MapGet("/rooms/{roomCode}", (HttpContext context, string roomCode) => GetRoom(context, roomCode, relayServer, relayHostOptions.HttpAdminTokenBytes));

		app.MapGet("/rooms/{roomCode}/public", (string roomCode) => GetPublicRoom(roomCode, relayServer))
		   .RequireRateLimiting(PublicRoomRateLimitPolicy);

		app.MapDelete("/rooms/{roomCode}", (HttpContext context, string roomCode) => DeleteRoom(context, roomCode, relayServer, relayHostOptions.HttpAdminTokenBytes));

		app.MapPatch("/rooms/{roomCode}", (HttpContext context, string roomCode, PatchRoomRequest patchRoomRequest) => PatchRoom(context, roomCode, patchRoomRequest, relayServer, relayHostOptions.HttpAdminTokenBytes));

		app.MapDelete("/rooms/{roomCode}/clients/{virtualClientId:int}", (HttpContext context, string roomCode, int virtualClientId) => KickClient(context, roomCode, virtualClientId, relayServer, relayHostOptions.HttpAdminTokenBytes));

		return app;
	}

	private static IResult CreateRoom(HttpContext context, CreateRoomRequest request, Server relayServer, byte[] adminTokenBytes)
	{
		if (!IsAuthorised(context, adminTokenBytes)) return Results.Unauthorized();

		if (request.MaximumClients <= 0) return Results.BadRequest($"{nameof(CreateRoomRequest.MaximumClients)} must be greater than 0.");

		if (Encoding.UTF8.GetByteCount(request.DisplayName) > 255) return Results.BadRequest($"{nameof(CreateRoomRequest.DisplayName)} must be at most 255 bytes when UTF-8 encoded.");

		bool created = relayServer.TryCreateReservedRoom(request.MaximumClients, request.DisplayName, request.IsPublic, request.Metadata, out RoomSnapshot? snapshot, out ErrorCode errorCode);

		if (created) return Results.Ok(snapshot);

		return errorCode switch
		{
			ErrorCode.RoomExists => Results.Conflict(errorCode.ToString()),

			_ => Results.BadRequest(errorCode.ToString()),
		};
	}

	private static IResult GetRooms(HttpContext context, Server relayServer, byte[] adminTokenBytes)
	{
		return !IsAuthorised(context, adminTokenBytes) ? Results.Unauthorized() : Results.Ok(relayServer.GetRoomSnapshots());
	}

	private static IResult GetRoom(HttpContext context, string roomCode, Server relayServer, byte[] adminTokenBytes)
	{
		if (!IsAuthorised(context, adminTokenBytes)) return Results.Unauthorized();

		RoomSnapshot? snapshot = relayServer.GetRoomSnapshot(roomCode);

		return snapshot is null ? Results.NotFound() : Results.Ok(snapshot);
	}

	private static IResult GetPublicRoom(string roomCode, Server relayServer)
	{
		if (!RoomCode.IsValid(roomCode)) return Results.NotFound();

		PublicRoomSnapshot? snapshot = relayServer.GetPublicRoomSnapshot(roomCode);

		return snapshot is null ? Results.NotFound() : Results.Ok(snapshot);
	}

	private static IResult DeleteRoom(HttpContext context, string roomCode, Server relayServer, byte[] adminTokenBytes)
	{
		string? bearerToken = GetBearerToken(context);

		if (!IsAuthorised(bearerToken, adminTokenBytes) && !HasRoomEditAccess(relayServer, roomCode, bearerToken)) return Results.Unauthorized();

		return relayServer.DeleteRoom(roomCode) ? Results.NoContent() : Results.NotFound();
	}

	private static string? GetBearerToken(HttpContext httpContext)
	{
		string? authorisationHeader = httpContext.Request.Headers.Authorization;

		if (string.IsNullOrWhiteSpace(authorisationHeader)) return null;

		return authorisationHeader.StartsWith("Bearer ", StringComparison.OrdinalIgnoreCase) ? authorisationHeader[7..] : null;
	}

	private static bool IsAuthorised(HttpContext httpContext, byte[] adminTokenBytes)
	{
		return IsAuthorised(GetBearerToken(httpContext), adminTokenBytes);
	}

	private static bool IsAuthorised(string? bearerToken, byte[] adminTokenBytes)
	{
		if (bearerToken is null) return false;

		int bearerTokenByteCount = Encoding.UTF8.GetByteCount(bearerToken);

		byte[] bearerTokenBytes = ArrayPool<byte>.Shared.Rent(bearerTokenByteCount);

		try
		{
			Encoding.UTF8.GetBytes(bearerToken, bearerTokenBytes);

			return CryptographicOperations.FixedTimeEquals(bearerTokenBytes.AsSpan(0, bearerTokenByteCount), adminTokenBytes);
		}
		finally
		{
			CryptographicOperations.ZeroMemory(bearerTokenBytes);

			ArrayPool<byte>.Shared.Return(bearerTokenBytes);
		}
	}

	private static bool HasRoomEditAccess(Server relayServer, string roomCode, string? bearerToken)
	{
		return !string.IsNullOrWhiteSpace(bearerToken) && relayServer.HasRoomHostToken(roomCode, bearerToken);
	}

	private static IResult PatchRoom(HttpContext context, string roomCode, PatchRoomRequest request, Server relayServer, byte[] adminTokenBytes)
	{
		string? bearerToken = GetBearerToken(context);

		if (!IsAuthorised(bearerToken, adminTokenBytes) && !HasRoomEditAccess(relayServer, roomCode, bearerToken)) return Results.Unauthorized();

		if (request.DisplayName is not null && Encoding.UTF8.GetByteCount(request.DisplayName) > 255) return Results.BadRequest($"{nameof(PatchRoomRequest.DisplayName)} must be at most 255 bytes when UTF-8 encoded.");

		string? metadataValidationError = ValidatePublicMetadata(request.MetadataToAdd, request.MetadataToRemove);

		if (metadataValidationError is not null) return Results.BadRequest(metadataValidationError);

		RoomSnapshot? snapshot = relayServer.PatchRoom(roomCode, request.DisplayName, request.IsPublic, request.MetadataToAdd, request.MetadataToRemove);

		return snapshot is null ? Results.NotFound() : Results.Ok(snapshot);
	}

	private static string? ValidatePublicMetadata(IReadOnlyDictionary<string, string>? metadataToAdd, IReadOnlyList<string>? metadataToRemove)
	{
		if (metadataToAdd is { Count: > MaximumMetadataEntries }) return $"At most {MaximumMetadataEntries} metadata entries may be added per request.";

		if (metadataToRemove is { Count: > MaximumMetadataEntries }) return $"At most {MaximumMetadataEntries} metadata entries may be removed per request.";

		int totalBytes = 0;

		if (metadataToAdd is not null)
		{
			foreach ((string key, string value) in metadataToAdd)
			{
				if (string.IsNullOrWhiteSpace(key) || !PublicMetadataKeys.Contains(key)) return $"Metadata key '{key}' is not allowed.";

				int keyBytes = Encoding.UTF8.GetByteCount(key);
				int valueBytes = Encoding.UTF8.GetByteCount(value ?? string.Empty);

				if (keyBytes > MaximumMetadataKeyBytes) return $"Metadata keys must be at most {MaximumMetadataKeyBytes} bytes when UTF-8 encoded.";

				if (valueBytes > MaximumMetadataValueBytes) return $"Metadata values must be at most {MaximumMetadataValueBytes} bytes when UTF-8 encoded.";

				totalBytes += keyBytes + valueBytes;
			}
		}

		if (totalBytes > MaximumMetadataTotalBytes) return $"Metadata must be at most {MaximumMetadataTotalBytes} bytes when UTF-8 encoded.";

		if (metadataToRemove is not null)
		{
			foreach (string key in metadataToRemove)
			{
				if (string.IsNullOrWhiteSpace(key) || !PublicMetadataKeys.Contains(key)) return $"Metadata key '{key}' is not allowed.";
			}
		}

		if (metadataToAdd is not null && metadataToAdd.TryGetValue("Users", out string? users) && !int.TryParse(users, NumberStyles.Integer, CultureInfo.InvariantCulture, out _)) return "Users must be an integer.";

		if (metadataToAdd is not null && metadataToAdd.TryGetValue("MaxMembers", out string? maxMembers) && !int.TryParse(maxMembers, NumberStyles.Integer, CultureInfo.InvariantCulture, out _)) return "MaxMembers must be an integer.";

		return null;
	}

	private static IResult KickClient(HttpContext context, string roomCode, int virtualClientId, Server relayServer, byte[] adminTokenBytes)
	{
		string? bearerToken = GetBearerToken(context);

		if (!IsAuthorised(bearerToken, adminTokenBytes) && !HasRoomEditAccess(relayServer, roomCode, bearerToken)) return Results.Unauthorized();

		return relayServer.TryKickClient(roomCode, virtualClientId) ? Results.NoContent() : Results.NotFound();
	}

	public sealed record CreateRoomRequest
	(
		ushort MaximumClients = 4096,
		string DisplayName = "",
		bool IsPublic = false,
		Dictionary<string, string>? Metadata = null
	);

	public sealed record PatchRoomRequest
	(
		string? DisplayName = null,
		bool? IsPublic = null,
		Dictionary<string, string>? MetadataToAdd = null,
		List<string>? MetadataToRemove = null
	);
}